Miner deploy rallypoint functionality without engine changes
